requirements.txt
flask==3.0.3
flask-cors==4.0.0
# CPU torch + torchvision for py3.11 linux x86_64
torch==2.2.2
torchvision==0.17.2
# Ultralytics version that runs YOLOv8/11 inference fine on torch 2.2.* 
ultralytics==8.2.103
# Matching libs that have wheels for py3.11
numpy<2.0
Pillow==10.3.0
opencv-python-headless==4.9.0.80
gunicorn==21.2.0
